Fixed mlpcolwise.py

+++ python/nerva/datasets_rowwise.py
@@ -141,7 +141,7 @@ class TorchDataLoader(object):
     A data loader with an interface similar to torch.utils.data.DataLoader.
     """
 
-    def __init__(self, Xdata: torch.Tensor, Tdata: torch.IntTensor, batch_size: int):
+    def __init__(self, Xdata: torch.Tensor, Tdata: torch.LongTensor, batch_size: int):
         """
         :param Xdata: a dataset
         :param Tdata: the targets for the dataset
